# Improve cuBLAS performance by using a memory pool