Previously, the cuda memory was allocated and freed as needed for each mat mul operation, which is very inefficient.
By using a memory pool, this is about 30-50% faster in my machine.

PR:
```
7B q4_0 3.59 seconds per pass - ETA 0.65 hours
7B f16  4.59 seconds per pass - ETA 0.83 hours
7B f32  6.68 seconds per pass - ETA 1.22 hours

 7B q4_0 llama_print_timings: prompt eval time =  5493.53 ms /   631 tokens (    8.71 ms per token)
13B q4_0 llama_print_timings: prompt eval time =  9003.90 ms /   631 tokens (   14.27 ms per token)
30B q4_0 llama_print_timings: prompt eval time = 19682.41 ms /   631 tokens (   31.19 ms per token)
```

Master:
```
7B q4_0 5.09 seconds per pass - ETA 0.93 hours

 7B q4_0 prompt eval time =  7840.48 ms /   631 tokens (   12.43 ms per token)
13B q4_0 prompt eval time = 13826.48 ms /   631 tokens (   21.91 ms per token)
```