@glinscott 2023-04-21
Very nice!  On, 30B `8.58 seconds per pass` now, `10.99 seconds per pass` before.