@slaren 2023-04-21
@dfyz I think it is very unlikely that a general purpose allocator like that will be faster than this, but if you want to give it a try please do. On the other hand, if it isn't compatible with HIPS and #1087 is merged, we would have to write a custom allocator anyway.

Good points with the synchronization issues, this is always harder than it looks. In my tests llama doesn't make any concurrent mat muls with BLAS, so my thought here was mostly about making it future proof, but if that is too complicated we may as well replace it with a spin lock with zero performance impact right now.