@SlyEcho 2023-04-21
HIP supports the same [async malloc/free operations](https://docs.amd.com/bundle/HIP-API-Guide-v5.4.1/page/a00200.html) and mempool stuff. Technically, it should be checked from the device's feature flags, but that would make our code more difficult.

When I was first getting hipBLAS working, I also tried hipMallocAsync and it really didn't make any difference. When I look at the profile, hipMalloc/Free is a tiny part of the matmul operation which is still dominated by hipMemcpy. Maybe if I try the changes in this PR it will reveal something different.

I think the only real impact would be to keep the big weight matrixes on the device permanently as much as it can fit.