@ggerganov 2023-04-21
@slaren 

If you manage to write an efficient `q4_x_q8` CUDA kernel you will probably achieve another factor of 2 or more speed-up.

Anyway - great work as always!

Regarding the race condition:
Without spending too much time on this, it looks to me it is not thread safe. @dfyz might have outlined a problematic scenario and explanation, but I don't want to spend too much time on analysing this. Just from my experience, you want be able to guarantee thread-safety here with just an atomic var.
I would recommend to remove the atomic pointers as we don't have a case where we do parallel BLAS calls.

On that note - would there be any sort of benefit theoretically from parallel GPU BLAS calls? I've tried on CPU and it doesn't help.

Can we add some parameter that specifies how much VRAM you want to be used at maximum?
For example, people with less VRAM would prefer to limit allocation to that certain point so they can run even the 65B models and still get benefits from the GPU