@SlyEcho 2023-04-21
There could be a lot of benefit from parallel GPU BLAS calls. For example we could use the [GemmStridedBatched](https://docs.nvidia.com/cuda/cublas/index.html#cublas-t-gemmstridedbatched) functions to compute the multiplication over the entire tensor instead of for-looping the outer dimensions (but I think llama.cpp doesn't need it anyway?).

But they also mention this:

> On certain problem sizes, it might be advantageous to make multiple calls to cublas<t>gemm in different CUDA streams, rather than use this API.

Currently, the computation (on AMD, at least, could be that Nvidia is better at this) there is just one row of operations for one matmul calculation.
![image](https://user-images.githubusercontent.com/795193/233673052-5207fb7d-98c0-4ab9-b6bb-8fe69d635421.png)

@slaren, do you think it would be worth having multiple cudaStreams, so that some operations could happen at the same time as others?

(In OpenCL (like clBLAS) every operation gives you event objects that other operations can wait on, so the entire ggml compute graph could be posted to the GPU command stream and let it sort out the dependencies and compute in parallel. But this is my guess if it actually works like that)