@slaren 2023-04-21
>I would recommend to remove the atomic pointers as we don't have a case where we do parallel BLAS calls.
Yes, I will replace the whole synchronization of the memory pool with a spin lock for now, just to make sure that it doesn't cause issues in the future.

>On that note - would there be any sort of benefit theoretically from parallel GPU BLAS calls?
I think so, currently the GPU sits idle while waiting for the data, and the PCIe bus is unused while computing the gemm, with more threads it would be possible to continue copying data for the next mat mul while a gemm is being computed in a different thread. Probably will require some changes to the cuda code, we would need to use a different stream per thread, but if it is not very hard to change ggml to allow this, I think it is worth giving it a try.

>Can we add some parameter that specifies how much VRAM you want to be used at maximum?
This PR does increase VRAM usage, I didn't test with 65B but with 30B it is still very low, below what any discrete GPU would have. If we implement multi-threaded BLAS mat muls that could change, though.