@slaren 2023-04-21
> @slaren, do you think it would be worth having multiple cudaStreams, so that some operations could happen at the same time as others?

Yes, that's what I was thinking.