@SlyEcho 2023-04-21
I think, maybe the simplest thing is two have two streams, for `src0` and `src1`. Then, before the Gemm, synchronize them.