@slaren 2023-04-21
> If you manage to write an efficient `q4_x_q8` CUDA kernel you will probably achieve another factor of 2 or more speed-up.

Here is an interesting article about how that could be done: https://siboehm.com/articles/22/CUDA-MMM