@slaren 2023-04-21
> Could you please clarify which commands (and which hardware) you used for benchmarking? I guess that the output with `ETA [...] hours` is the output of `perplexity` command, but I'm less sure about the `prompt eval time` lines.

Yes, the perplexity times is just from running perplexity with default params and the standard wiki test file. The only additional parameter is `-t 12`, which is what works well for me.

The prompt eval times are just from running `main` with a large prompt file and `-b 512 -n 1 -t 12`.