@slaren 2023-04-21
The synchronization issues should be resolved, and with a 30B q4_0 model I see a VRAM usage of ~2 GB, so I don't expect that it will be a problem even with 65B models on very low VRAM GPUs.

We could do better if we were able to predict the allocations, currently a single pass will make these allocations:
```
ggml_compute_forward_mul_mat_q_f32: (6656 x 6656) x (512 x 6656) => (512 x 6656)
cuda allocated 169 MB: total: 169 MB
cuda allocated 13 MB: total: 182 MB
cuda allocated 13 MB: total: 195 MB
cuda allocated 26 MB: total: 221 MB

ggml_compute_forward_mul_mat_q_f32: (17920 x 6656) x (512 x 6656) => (512 x 17920)
cuda allocated 455 MB: total: 676 MB
cuda allocated 35 MB: total: 711 MB
cuda allocated 71 MB: total: 782 MB

ggml_compute_forward_mul_mat_q_f32: (32000 x 6656) x (512 x 6656) => (512 x 32000)
cuda allocated 812 MB: total: 1595 MB
cuda allocated 126 MB: total: 1721 MB
```

In principle, we could allocate only the memory for the (32000 x 6656) x (512 x 6656) mat mul and reuse these buffers for the smaller mat muls. For now this isn't an issue, but that may change if we implement multi-threaded mat muls.