@Dampfinchen 2023-04-27
@slaren  I wonder if tensor cores could provide another huge speedup when optimizing for them. Judging by how GPU architectures with similar raw performance (one with and one without tensor cores) perform in llama's cuBLAS implementation right now, tensor cores are either not in use at all currently or used inefficiently. I've found the following documents about that matter, maybe you can look into them:

https://forums.developer.nvidia.com/t/turing-arch-int4-ops-with-tensor-cores/66656
https://docs.nvidia.com/cuda/cublas/index.html?highlight=tensor%20cores#tensor-core-usage

Thanks for your great work. I'm looking forward to your next innovation for cuBLAS!