@slaren 2023-04-27
cuBLAS already uses tensors cores, see https://docs.nvidia.com/cuda/cublas/index.html#tensor-core-usage

Unfortunately, it seems that NVIDIA intends to deprecate INT4 support in tensor cores, and it has already been removed in sm_90 (Hopper / H100), but we may still benefit from it with RTX 20 to 40 series cards if we write our own mat mul kernel in the future.